refactor: Fund Page Widget 분리
